Move leave game button to header bar. The leave game button at the bottom was overlapping with the footer bar.
